Moved VirtualMachine to own file (#12)
